Course22 - JPA relations added

+++ src/main/java/ro/fasttrackit/countriesapp/exception/ApiError.java
@@ -5,6 +5,6 @@ import lombok.Builder;
 @Builder
 public record ApiError(
         String message,
-        int entityId
+        long entityId
 ) {
 }
