+++ src/main/java/ro/fasttrackit/countriesapp/exception/ResourceNotFoundException.java
@@ -4,9 +4,9 @@ import lombok.Getter;
 
 @Getter
 public class ResourceNotFoundException extends RuntimeException {
-    private final int entityId;
+    private final long entityId;
 
-    public ResourceNotFoundException(String message, int entityId) {
+    public ResourceNotFoundException(String message, long entityId) {
         super(message);
         this.entityId = entityId;
     }
